Sets the sql max idle connections to 1
turns out without making a call `sqlConn.SetMaxIdleConns' the default idle
connections will be set to 0, i.e. no idle connections will be kept.
this results in a high number of connections closed and left in the
`TIME_WAIT' state which will eventually cause the process to run out of
fd

[#113854837]

Signed-off-by: John Shahid <jshahid@pivotal.io>