Set the buffer on receiving channel for BBS events
BBS used to drop the route emitter subscriber defined in tests due to
slow consumer. Add a buffer to the channel to allow the messages to be
cached on a client side. The size of the buffer is the average total
number of events that is being sent in a test. Seems like that does not
consume much memory on diego benchmarks VM.

[#168790356](https://www.pivotaltracker.com/story/show/168790356)